requirements.txt
yt-dlp
python-telegram-bot==20.7
tinydb
